// src/types.ts
export interface Measurement {
  measured_at: string;
  weight: number;
  bmi: number;
  body_fat: number;
  fat_free_weight: number;
  subcutaneous_fat: number;
  visceral_fat: number;
  body_water: number;
  muscle_mass: number;
  skeletal_muscles: number;
  bone_mass: number;
  protein: number;
  bmr: number;
  metabolic_age: number;
}